Extract functions into files and disable modals
