+++ assets/js/common/index.js
@@ -2,6 +2,7 @@ import forEach from 'lodash/forEach';
 import fitvids from 'fitvids';
 import tabs from './tabs';
 import forms from './forms';
+import session from './session';
 
 /**
  * Common modules that are run across the site.
@@ -19,4 +20,5 @@ export const init = () => {
 
     tabs.init();
     forms.init();
+    session.init();
 };
